+++ src/main/resources/assets/mwc/models/item/ceiling_lamp.json
@@ -1,7 +1,7 @@
 {
     "parent": "builtin/generated",
     "textures": {
-        "layer0": "mwc:items/props/ceiling_lamp"
+        "layer0": "mwc:items/props/lights/ceiling_lamp"
     },
     "display": {
 		"thirdperson_righthand": {
