+++ src/main/resources/assets/mwc/models/item/ceiling_lamp_alt.json
@@ -1,7 +1,7 @@
 {
     "parent": "builtin/generated",
     "textures": {
-        "layer0": "mwc:items/props/ceiling_lamp_alt"
+        "layer0": "mwc:items/props/lights/ceiling_lamp_alt"
     },
     "display": {
 		"thirdperson_righthand": {
