+++ src/main/resources/assets/mwc/models/item/shelf_left.json
@@ -1,7 +1,7 @@
 {
     "parent": "builtin/generated",
     "textures": {
-        "layer0": "mwc:items/props/shelf_left"
+        "layer0": "mwc:items/props/shelves/shelf_left"
     },
     "display": {
 		"thirdperson_righthand": {
