+++ src/main/resources/assets/mwc/models/item/shelf_right.json
@@ -1,7 +1,7 @@
 {
     "parent": "builtin/generated",
     "textures": {
-        "layer0": "mwc:items/props/shelf_right"
+        "layer0": "mwc:items/props/shelves/shelf_right"
     },
     "display": {
 		"thirdperson_righthand": {
