+++ src/main/resources/assets/mwc/models/item/shelf_top.json
@@ -1,7 +1,7 @@
 {
     "parent": "builtin/generated",
     "textures": {
-        "layer0": "mwc:items/props/shelf_top"
+        "layer0": "mwc:items/props/shelves/shelf_top"
     },
     "display": {
 		"thirdperson_righthand": {
