+++ src/main/resources/assets/mwc/models/item/white_desk_left.json
@@ -1,7 +1,7 @@
 {
     "parent": "builtin/generated",
     "textures": {
-        "layer0": "mwc:items/props/white_desk_left"
+        "layer0": "mwc:items/props/desks/white_desk_left"
     },
     "display": {
 		"thirdperson_righthand": {
