+++ src/main/resources/assets/mwc/models/item/white_desk_middle.json
@@ -1,7 +1,7 @@
 {
     "parent": "builtin/generated",
     "textures": {
-        "layer0": "mwc:items/props/white_desk_middle"
+        "layer0": "mwc:items/props/desks/white_desk_middle"
     },
     "display": {
 		"thirdperson_righthand": {
