+++ src/main/resources/assets/mwc/models/item/white_desk_right.json
@@ -1,7 +1,7 @@
 {
     "parent": "builtin/generated",
     "textures": {
-        "layer0": "mwc:items/props/white_desk_right"
+        "layer0": "mwc:items/props/desks/white_desk_right"
     },
     "display": {
 		"thirdperson_righthand": {
