+++ src/main/resources/assets/mwc/textures/items/props/lights/ceiling_lamp_alt.png
