+++ src/main/resources/assets/mwc/textures/items/props/lights/floodlight.png
