+++ src/main/resources/assets/mwc/textures/models/props/lights/duelfloodlight.png
